+++ src/game/BattleGround/BattleGroundMgr.cpp
@@ -1414,9 +1414,9 @@ void BattleGroundMgr::LoadBattleEventIndexes()
     events.event1 = BG_EVENT_NONE;
     events.event2 = BG_EVENT_NONE;
     m_GameObjectBattleEventIndexMap.clear();             // need for reload case
-    m_GameObjectBattleEventIndexMap[-1] = events;
+    m_GameObjectBattleEventIndexMap[static_cast<uint32>(-1)] = events;
     m_CreatureBattleEventIndexMap.clear();               // need for reload case
-    m_CreatureBattleEventIndexMap[-1] = events;
+    m_CreatureBattleEventIndexMap[static_cast<uint32>(-1)] = events;
 
     uint32 count = 0;
