+++ src/game/MotionGenerators/HomeMovementGenerator.cpp
@@ -57,7 +57,7 @@ void HomeMovementGenerator<Creature>::_setTargetLocation(Creature& owner)
     init.Launch();
 
     arrived = false;
-    owner.clearUnitState(UNIT_STAT_ALL_DYN_STATES);
+    owner.clearUnitState(static_cast<uint32>(UNIT_STAT_ALL_DYN_STATES));
 }
 
 bool HomeMovementGenerator<Creature>::Update(Creature& owner, const uint32& /*time_diff*/)
